Indépendance des réglages et boutons et commentaires + ajout réglages au builder
