+++ Idées LOGO/#Index_logos_builder.csv
@@ -1,0 +1,12 @@
+#Nom du fichier du logo, nom du logo
+logo builder 1.jpg
+logo builder 2.jpg
+logo builder 3.jpg,defaut
+logo builder 4.jpg
+logo builder 5.jpg,visage
+logo builder 6.jpg,militaire
+logo builder 7.jpg
+logo builder 8.jpg
+logo builder 9.jpg
+logo builder 10.jpg
+logo builder 11.jpg
