+++ Idées LOGO/#Index_logos_parcoureur.csv
@@ -1,0 +1,14 @@
+#Nom du fichier du logo, nom du logo
+logo 1.jpg
+logo 2.jpg,defaut
+logo 3.jpg,insta
+logo 4.jpg
+logo 5.jpg,water
+logo 6.jpg
+logo 7.jpg
+logo 8.jpg
+logo 9.jpg
+logo 10.jpg
+logo 11.jpg
+logo 12.jpg
+logo 13.jpg
