--- Parametres_defaut.csv
@@ -1,24 +1,0 @@
-#Nom du parametre, valeur du parametre
-type deplacement initial,Lisse
-lab alea x,10
-lab alea y,5
-lab alea entree x,0
-lab alea entree y,0
-lab alea alerte mauvaise entree,1
-nb colones min,3
-nb colones max,50
-nb lignes min,3
-nb lignes max,35
-initial color mode,black
-decoupe du deplacement,5
-vitesse deplacement,45
-question confirmation lab suivant,0
-question confirmation lab precedent,0
-question confirmation recomencer lab,1
-question confirmation passage niveau 2,0
-question confirmation passage niveau 3,0
-question confirmation passage niveau 4,0
-question confirmation passage niveau max,1
-color mode hard ready,green
-colors mode hard moving,blue,purple,grey
-color mode hard impossible,red
